Skin Prophet (WIP)
Skin Prophet additions
- Summon Fire Sphere: spawns a Fire Sphere enemy that explodes into fire after a short period of time (*)
- "Powerful short spell": a close range aoe that can be casted as long as there is one Lit Candle nearby
- when the Skin Prophet dies, it changes to the second phase of the encounter (**)

Skin Prophet (second phase)
- "Light attack": deals double damage, however damage is nullified if the player has a fire feather in their inventory (***)
- Summon Fire Sphere (second phase): spawns a Fire Sphere that explodes quicker than before, deals more dmg
- "Scatter": casts a random aoe around the player, deals damage after 5 s

The following is WIP
(*) death_function is very buggy for some reason, it only works with a spell named as "death_conflagration". All death_effects in this PR use death_conflagration to debug, data for the final spells is added but is not used because of the bug
(**) not working for mon_prophet_2, but it works for mon_halfworm
(***) the item is not implemented yet, has to be dropped during the encounter

A custom "Survivor" profession was added with debug spells. An edited "mon_zombie" was added for debugging purposes too